fix: util usage

+++ decoder/gym.go
@@ -181,7 +181,7 @@ func (gym *Gym) updateGymFromFort(fortData *pogo.PokemonFortProto, cellId uint64
 		TempEvolutionFinishMs int64 `json:"temp_evolution_finish_ms,omitempty"`
 		Alignment             int   `json:"alignment,omitempty"`
 		Badge                 int   `json:"badge,omitempty"`
-		Background            int   `json:"background,omitempty"`
+		Background            int64 `json:"background,omitempty"`
 	}
 	gym.Id = fortData.FortId
 	gym.Lat = fortData.Latitude  //fmt.Sprintf("%f", fortData.Latitude)
@@ -331,7 +331,7 @@ func (gym *Gym) updateGymFromGymInfoOutProto(gymData *pogo.GymGetInfoOutProto) *
 		TempEvolutionFinishMs int64              `json:"temp_evolution_finish_ms,omitempty"`
 		Alignment             int                `json:"alignment,omitempty"`
 		Badge                 int                `json:"badge,omitempty"`
-		Background            int                `json:"background,omitempty"`
+		Background            int64              `json:"background,omitempty"`
 		DeployedMs            int64              `json:"deployed_ms,omitempty"`
 		DeployedTime          int64              `json:"deployed_time,omitempty"`
 		BattlesWon            int32              `json:"battles_won"`
